Make test suites aware of test groups
In this architecture, a test suite is composed of one or more test groups.
Each test belongs to a test group. The test group defines the setup and
teardown procedures that are run for a test.

The interface should provide that individual tests can be run without defining
a test group if it is not needed. This can be handled by adding orphan tests
to a default test group that has setup/teardown functions that take no action.