Remove igorutils submodule from igorunit repo
IgorUnit depends on igorutils, so I initially thought that it would be easiest
to distribute the igorutils from with IgorUnit. Howver, only one copy of a
procedure file is available to Igor at a time. In order for more than one
package to use igorutils, igorutils must be installed independently. The
igorutils copy inside of IgorUnit was hiding other, more recent versions of
igorutils so I am deleting it from IgorUnit. Note that IgorUnit still requires
igorutils!